commit20240821

+++ file/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render
 from django.http import JsonResponse
 from .models import UploadFile, FileInfo
+from triples.models import Triples
 import json
 import PyPDF2
 
@@ -26,12 +27,12 @@ def file_info(requst, pk):
 
 def file(reqeust, pk):
 	try:
-		file_info = FileInfo.objects.get(id=pk)
+		info = FileInfo.objects.get(id=pk)
 	
 	except FileInfo.DoesNotExist:
 		return returnJson([], 0, 404)
 	
-	file = UploadFile.objects.filter(file_info=file_info).first()
+	file = UploadFile.objects.filter(info=info).first()
 
 	if file is None:
 		return returnJson()
@@ -48,6 +49,13 @@ def create_file_info(request):
 		file_info = FileInfo.objects.create(name=data['name'], created_by=data['username'], json_created=data['json_created'])
 		file_info.save()
 
+		for triple in data['json_created']:
+			Triples.objects.create(
+                entity1=triple['entity1'],
+                relationship=triple['relationship'],
+                entity2=triple['entity2'],
+            )
+
 		return returnJson([dict(file_info.body())])
 	
 	return returnJson([], 0, 400)
@@ -58,18 +66,25 @@ def create_file(request, pk):
 	except FileInfo.DoesNotExist:
 		return returnJson([], 0, 404)
 	
-	# text = extract_text_from_pdf(request.FILES.get('file'))
 	file = UploadFile.objects.create(info=file_info, file=request.FILES.get('file'))
 
 	return returnJson([dict(file.body())])
 
 def delete_file(request, pk):
+	username = request.user.username
 	file_info = FileInfo.objects.get(id=pk)
 
 	if request.method == 'DELETE':
-			file_info.delete()
-			files = UploadFile.objects.all()
-			return returnJson([dict(file.body()) for file in files])
+			print(file_info.created_by + username)
+			if file_info.created_by.strip() == username.strip():
+				file = UploadFile.objects.filter(info=file_info).first()
+				file.delete()
+				file_info.delete()
+				files = UploadFile.objects.all()
+				return returnJson([dict(file.body()) for file in files])
+			
+			else:
+				return returnJson([], 0, 400)
 
 def extract_text_from_pdf(request):
     file = request.FILES.get('file')
